consolidate libmessages into libhal
