--- HAL/Messages/PbMsgsConfig.cmake.in
@@ -1,10 +1,0 @@
-# Compute paths
-get_filename_component( PROJECT_CMAKE_DIR "${CMAKE_CURRENT_LIST_FILE}" PATH )
-SET( @PROJECT_NAME@_INCLUDE_DIRS "@EXPORT_LIB_INC_DIR@;@USER_INC@" )
-
-# Library dependencies (contains definitions for IMPORTED targets)
-if( NOT TARGET @LIBRARY_NAME@ AND NOT @PROJECT_NAME@_BINARY_DIR )
-  include( "${PROJECT_CMAKE_DIR}/@PROJECT_NAME@Targets.cmake" )
-endif()
-
-SET( @PROJECT_NAME@_LIBRARIES @LIBRARY_NAME@ )
