--- HAL/Messages/PbMsgsConfigVersion.cmake.in
@@ -1,16 +1,0 @@
-set(PACKAGE_VERSION "@VERSION@")
-
-# Check build type is valid
-if( "${ANDROID}" STREQUAL "@ANDROID@" )
-    # Check whether the requested PACKAGE_FIND_VERSION is compatible
-    if("${PACKAGE_VERSION}" VERSION_LESS "${PACKAGE_FIND_VERSION}")
-      set(PACKAGE_VERSION_COMPATIBLE FALSE)
-    else()
-      set(PACKAGE_VERSION_COMPATIBLE TRUE)
-      if ("${PACKAGE_VERSION}" VERSION_EQUAL "${PACKAGE_FIND_VERSION}")
-        set(PACKAGE_VERSION_EXACT TRUE)
-      endif()
-    endif()
-else()
-    set(PACKAGE_VERSION_COMPATIBLE FALSE)
-endif()
