--- HAL/Messages/config.h.in
@@ -1,3 +1,0 @@
-#pragma once
-
-#define PBMSGS_VERSION @VERSION_MAJOR@
